Move core import classes into `lib-import-export`
