+++ src/DropTarget.cpp
@@ -18,7 +18,7 @@
 #include "ProjectFileManager.h"
 #include "ProjectWindow.h"
 #include "TrackPanel.h"
-#include "import/Import.h"
+#include "Import.h"
 #include "import/ImportMIDI.h"
 
 #if wxUSE_DRAG_AND_DROP
