+++ src/ProjectFileManager.cpp
@@ -44,9 +44,9 @@ Paul Licameli split from AudacityProject.cpp
 #include "WaveClip.h"
 #include "wxFileNameWrapper.h"
 #include "Export.h"
-#include "import/Import.h"
-#include "import/ImportProgressListener.h"
-#include "import/ImportPlugin.h"
+#include "Import.h"
+#include "ImportProgressListener.h"
+#include "ImportPlugin.h"
 #include "import/ImportMIDI.h"
 #include "import/ImportStreamDialog.h"
 #include "toolbars/SelectionBar.h"
