+++ src/prefs/ExtImportPrefs.h
@@ -16,7 +16,7 @@
 
 #include "PrefsPanel.h"
 
-#include "../import/ImportPlugin.h"
+#include "ImportPlugin.h"
 
 class wxButton;
 class wxGridEvent;
